split up header files, and change mock client
