+++ libs/ipc/inc/public/tfc/ipc.hpp
@@ -3,6 +3,7 @@
 #include <system_error>
 
 #include <tfc/ipc/details/dbus_server_iface.hpp>
+#include <tfc/ipc/details/dbus_server_iface_mock.hpp>
 #include <tfc/ipc/details/impl.hpp>
 
 namespace tfc::ipc {
@@ -72,6 +73,10 @@ public:
   auto operator=(slot const&) -> slot& = delete;
 
   [[nodiscard]] auto value() const noexcept { return slot_->get(); }  // todo: value() or get()
+
+  auto something() -> std::string {
+    return "something";
+  }
 
 private:
   static constexpr std::string_view self_name{ slot_tag };
